@@ -1,7 +1,7 @@
 name: flutter_automated_tests
 
 environment:
-  sdk: ">=2.17.0-0 <4.0.0"
+  sdk: '>=2.19.0-0 <4.0.0'
 
 dependencies:
   flutter:
