@@ -2,7 +2,7 @@ name: complex_layout
 description: A benchmark of a relatively complex layout.
 
 environment:
-  sdk: ">=2.17.0-0 <4.0.0"
+  sdk: '>=2.19.0-0 <4.0.0'
 
 dependencies:
   flutter:
